wan-runpod-worker/requirements.txt
runpod>=1.7.0
torch>=2.4.0
torchvision
diffusers>=0.33.0
transformers>=4.46.0
accelerate>=0.34.0
huggingface_hub>=0.24.0
imageio[ffmpeg]>=2.34.0
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
sentencepiece
ftfy
